pyproject.toml
[tool.poetry]
name = "backend-vitrine-patrimonio"
version = "0.1.0"
description = ""
authors = ["ovictorhugo"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
psycopg2-binary = "^2.9.9"
flask = "^3.0.3"
flask-cors = "^5.0.0"
pandas = "^2.2.2"
pydantic = {extras = ["email"], version = "^2.8.2"}
python-dotenv = "^1.0.1"
requests = "^2.32.3"
secure-smtplib = "^0.1.1"
gunicorn = "^23.0.0"
xlrd = "^2.0.1"
openpyxl = "^3.1.5"
pyxlsb = "^1.0.10"
python-calamine = "^0.3.1"
firebase-admin = "^6.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
